opt: add rule to extract filters from OR clauses in filters
This PR adds a rule that is able to extract a filter from an
OR clause. It may then be possible to push the new filter item
down the expression tree. The following is an example of this
transformation:
```
(x = 1 AND y = "one") OR (x = 2 AND z = '10'::VARBIT)
=>
(x = 1 AND y = "one") OR (x = 2 AND z = '10'::VARBIT)
AND (x = 1 OR x = 2)
```

Release note: None