opt: separate UDF callsite from its definition
This patch splits the old `UDFExpr` into two new structs: `UDFCallExpr`
and `UDFDefinition`. The definition contains the body statements of the
UDF, while the call contains other information such as the arguments and
return type. This refactor is necessary to allow recursive UDFs, since a
recursive UDF needs to be able to refer back to itself before its
body statements have been built.

When printing a recursive function verbosely, the `ExprFmtCtx` keeps track
of which UDFs have already been seen while printing a UDF body, so that the
body is not printed for recursive calls. This patch does not allow users to
define recursive UDFs, but will allow recursive UDFs to be defined and
executed internally.

Release note: None